+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Table Readers moved to geopackage-core
 * Automatically enable foreign key checks for new connections without preexisting violations
 * GeoPackage table foreign key check
+* OGC OpenAPI Features download to GeoPackage support
 * Android Q fix for SQL Utils quote wrap
 * gradle plugin updated to 3.4.1
 * gradle version 5.1.1